+++ src/components/patient/MiniCard.jsx
